move RedisNull to be internal to BinarySerializer
